Remove enable-tftp from default dnsmasq config
Ironic-inspector is a secondary service to Ironic which can additionally
use a tertiary instance of dnsmasq to support hardware discovery and
completely disjointed from OpenStack based introspection of hardware.

This third service, dnsmasq, in practical use cases has it's configuration
entirely overwritten as it contains specific details to the IP addressing
of the local networks. This configuration which is overwritten does not
enable a tftp server. That doesn't mean it couldn't be used that way, but
pointers to another, primary, tftp server is more logical of a secenario
to support.

As such, it seems logical to remove the reference to enable-tftp as it is
not used in reality.

Change-Id: I49de85626538712174b9ed8b4780a99bdc96483c